refactor(dashboard): give PolicyForm and the policy model their own modules
A story importing the form from `RoutingPage` dragged the table, the
confirm dialog, `RouterReadiness` and the page's whole hook graph into
every render, and the 900-line dialog body now has a catalog entry of its
own. So `features/routing/PolicyForm.tsx` holds the form and the six
helpers only it uses.

Eleven types, constants and derivations are read by both sides, so they
go to `features/routing/policyModel.ts` rather than being imported back
out of the page: a component module reaching into a page module is the
dependency this split exists to keep pointing one way. Two of the eleven
(`candidatesOf`, `normalizedBackend`) are there because the shared
derivations call them, which a first pass at the seam missed.

A pure move. Verified by inventory rather than by eye: 26 top-level
declarations before, 26 after, none lost and none invented, 7 left on the
page, 11 in the model, 8 in the form.

`foundation.test.ts` caught the one thing a move like this can break
quietly. The caption-size exemption is a named ruling about consequence
notes beside a button, and four of its five sites moved, so the ruling
now names both files rather than the visuals changing inside a refactor.
