+++ web/src/styles/foundation.test.ts
@@ -1092,10 +1092,13 @@ describe("content text wears a type role", () => {
       "the centered next-step prose, page-referent",
     ],
     // The named edge: a note about what an action will do, in the row with
-    // the button, rather than about how to operate a control.
+    // the button, rather than about how to operate a control. Two entries for
+    // one ruling, because the form moved to its own module and took four of the
+    // five notes with it; the page kept the candidate cap.
+    ["features/routing/PolicyForm.tsx", "consequence notes in the button row"],
     [
       "features/routing/RoutingPage.tsx",
-      "consequence notes in the button row, and the candidate cap",
+      "the candidate cap, a consequence note in the button row",
     ],
     ["features/usage/ShareDialog.tsx", "a notice in the dialog's button row"],
   ]
